Move prompt cache tests to a feature-centric file and cover gaps
