+++ tests/models/test_openai_responses.py
@@ -6,7 +6,6 @@ from contextlib import asynccontextmanager
 from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Literal, cast
-from unittest.mock import AsyncMock
 
 import pytest
 from pydantic import BaseModel
@@ -16,7 +15,6 @@ from vcr.cassette import Cassette
 from pydantic_ai import (
     BinaryContent,
     BinaryImage,
-    CachePoint,
     CompactionPart,
     DocumentUrl,
     FilePart,
@@ -50,7 +48,6 @@ from pydantic_ai.agent import Agent
 from pydantic_ai.capabilities import NativeTool
 from pydantic_ai.direct import model_request as direct_model_request
 from pydantic_ai.exceptions import ContentFilterError, ModelHTTPError, ModelRetry
-from pydantic_ai.messages import ModelMessagesTypeAdapter
 from pydantic_ai.models import ModelRequestParameters
 from pydantic_ai.native_tools import CodeExecutionTool, FileSearchTool, ImageAspectRatio, MCPServerTool, WebSearchTool
 from pydantic_ai.output import NativeOutput, PromptedOutput, TextOutput, ToolOutput
@@ -78,11 +75,10 @@ with try_import() as imports_successful:
     )
     from openai.types.responses.response_refusal_delta_event import ResponseRefusalDeltaEvent
     from openai.types.responses.response_refusal_done_event import ResponseRefusalDoneEvent
-    from openai.types.responses.response_usage import InputTokensDetails, OutputTokensDetails, ResponseUsage
+    from openai.types.responses.response_usage import ResponseUsage
 
     from pydantic_ai.models.anthropic import AnthropicModel, AnthropicModelSettings
     from pydantic_ai.models.openai import (
-        OpenAIPromptCacheOptions,
         OpenAIResponsesModel,
         OpenAIResponsesModelSettings,
         _resolve_openai_image_generation_size,  # pyright: ignore[reportPrivateUsage]
@@ -168,379 +164,6 @@ async def test_openai_responses_image_detail_vendor_metadata(allow_model_request
     ]
     assert image_parts
     assert all(part['detail'] == 'high' for part in image_parts)
-
-
-# These adapter-level tests intentionally use mocked SDK clients: they assert exact request kwargs,
-# provider-specific omission, and pre-request guards that recordings cannot reliably pin.
-@pytest.mark.parametrize('provider_name', ['openai', 'openrouter'])
-async def test_openai_responses_cache_point_and_options(
-    allow_model_requests: None, provider_name: Literal['openai', 'openrouter']
-):
-    c = response_message(
-        [
-            ResponseOutputMessage(
-                id='output-1',
-                content=cast(list[Content], [ResponseOutputText(text='done', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ]
-    )
-    mock_client = MockOpenAIResponses.create_mock(c)
-    if provider_name == 'openai':
-        model = OpenAIResponsesModel('gpt-5.6-sol', provider=OpenAIProvider(openai_client=mock_client))
-    else:
-        model = OpenAIResponsesModel('openai/gpt-5.6-sol', provider=OpenRouterProvider(openai_client=mock_client))
-    settings = OpenAIResponsesModelSettings(openai_prompt_cache_options={'mode': 'explicit', 'ttl': '30m'})
-
-    result = await Agent(model, model_settings=settings).run(
-        ['Stable reference material.', CachePoint(ttl='1h'), 'Use the reference.']
-    )
-
-    assert result.output == 'done'
-    request = get_mock_responses_kwargs(mock_client)[0]
-    assert request['prompt_cache_options'] == {'mode': 'explicit', 'ttl': '30m'}
-    assert request['input'] == snapshot(
-        [
-            {
-                'role': 'user',
-                'content': [
-                    {
-                        'type': 'input_text',
-                        'text': 'Stable reference material.',
-                        'prompt_cache_breakpoint': {'mode': 'explicit'},
-                    },
-                    {'type': 'input_text', 'text': 'Use the reference.'},
-                ],
-            }
-        ]
-    )
-
-
-@pytest.mark.parametrize('mode', ['implicit', 'explicit'])
-async def test_openai_responses_prompt_cache_options_without_marker(
-    allow_model_requests: None, mode: Literal['implicit', 'explicit']
-):
-    """Request-wide cache options are independent of explicit breakpoint markers."""
-    response = response_message(
-        [
-            ResponseOutputMessage(
-                id='output-1',
-                content=cast(list[Content], [ResponseOutputText(text='done', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ]
-    )
-    mock_client = MockOpenAIResponses.create_mock(response)
-    model = OpenAIResponsesModel('gpt-5.6-sol', provider=OpenAIProvider(openai_client=mock_client))
-    settings = OpenAIResponsesModelSettings(openai_prompt_cache_options={'mode': mode})
-
-    await Agent(model, model_settings=settings).run('No explicit marker.')
-
-    request = get_mock_responses_kwargs(mock_client)[0]
-    assert request['prompt_cache_options'] == {'mode': mode}
-    assert request['input'] == [{'role': 'user', 'content': 'No explicit marker.'}]
-
-
-async def test_openai_responses_prompt_cache_options_without_breakpoint_support(allow_model_requests: None):
-    """Request-level cache options do not require explicit breakpoint support."""
-    response = response_message(
-        [
-            ResponseOutputMessage(
-                id='output-1',
-                content=cast(list[Content], [ResponseOutputText(text='done', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ]
-    )
-    mock_client = MockOpenAIResponses.create_mock(response)
-    model = OpenAIResponsesModel(
-        'custom-model',
-        provider=OpenAIProvider(openai_client=mock_client),
-        profile=OpenAIModelProfile(openai_responses_prompt_cache_supported_modes=frozenset({'implicit'})),
-    )
-
-    await Agent(
-        model,
-        model_settings=OpenAIResponsesModelSettings(openai_prompt_cache_options={'mode': 'implicit'}),
-    ).run('No explicit marker.')
-
-    request = get_mock_responses_kwargs(mock_client)[0]
-    assert request['prompt_cache_options'] == {'mode': 'implicit'}
-
-
-@pytest.mark.parametrize(
-    'options',
-    [
-        {'mode': 'implicit'},
-        {'ttl': '30m'},
-    ],
-)
-async def test_openrouter_responses_unsupported_prompt_cache_options_omitted(
-    allow_model_requests: None,
-    options: 'OpenAIPromptCacheOptions',
-):
-    """OpenRouter accepts request-level options only when `mode='explicit'`."""
-    response = response_message(
-        [
-            ResponseOutputMessage(
-                id='output-1',
-                content=cast(list[Content], [ResponseOutputText(text='done', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ]
-    )
-    mock_client = MockOpenAIResponses.create_mock(response)
-    model = OpenAIResponsesModel('openai/gpt-5.6-sol', provider=OpenRouterProvider(openai_client=mock_client))
-
-    await Agent(model, model_settings=OpenAIResponsesModelSettings(openai_prompt_cache_options=options)).run(
-        'No explicit marker.'
-    )
-
-    request = get_mock_responses_kwargs(mock_client)[0]
-    assert 'prompt_cache_options' not in request
-
-
-async def test_openai_responses_cache_point_history_prefix_stability(allow_model_requests: None):
-    """A serialized history preserves the cacheable prefix and its breakpoint across turns."""
-    response = response_message(
-        [
-            ResponseOutputMessage(
-                id='output-1',
-                content=cast(list[Content], [ResponseOutputText(text='done', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ]
-    )
-    mock_client = MockOpenAIResponses.create_mock([response, response])
-    model = OpenAIResponsesModel('gpt-5.6-sol', provider=OpenAIProvider(openai_client=mock_client))
-    agent = Agent(model)
-
-    first_result = await agent.run(['Stable context.', CachePoint(), 'First question.'])
-    history = ModelMessagesTypeAdapter.validate_json(ModelMessagesTypeAdapter.dump_json(first_result.all_messages()))
-    await agent.run('Follow-up question.', message_history=history)
-
-    first_request, second_request = get_mock_responses_kwargs(mock_client)
-    first_input = cast(list[dict[str, Any]], first_request['input'])
-    second_input = cast(list[dict[str, Any]], second_request['input'])
-    assert second_input[0] == first_input[0]
-    assert second_input[0] == snapshot(
-        {
-            'role': 'user',
-            'content': [
-                {
-                    'type': 'input_text',
-                    'text': 'Stable context.',
-                    'prompt_cache_breakpoint': {'mode': 'explicit'},
-                },
-                {'type': 'input_text', 'text': 'First question.'},
-            ],
-        }
-    )
-    assert second_input[-1] == {'role': 'user', 'content': 'Follow-up question.'}
-
-
-async def test_openai_responses_image_cache_point(allow_model_requests: None):
-    """Not a VCR test: this pins OpenAI's image-block breakpoint translation."""
-    c = response_message(
-        [
-            ResponseOutputMessage(
-                id='output-1',
-                content=cast(list[Content], [ResponseOutputText(text='done', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ]
-    )
-    mock_client = MockOpenAIResponses.create_mock(c)
-    model = OpenAIResponsesModel('gpt-5.6-sol', provider=OpenAIProvider(openai_client=mock_client))
-
-    await Agent(model).run([ImageUrl('https://example.com/reference.png'), CachePoint(), 'Describe the reference.'])
-
-    assert get_mock_responses_kwargs(mock_client)[0]['input'] == snapshot(
-        [
-            {
-                'role': 'user',
-                'content': [
-                    {
-                        'type': 'input_image',
-                        'detail': 'auto',
-                        'image_url': 'https://example.com/reference.png',
-                        'prompt_cache_breakpoint': {'mode': 'explicit'},
-                    },
-                    {'type': 'input_text', 'text': 'Describe the reference.'},
-                ],
-            }
-        ]
-    )
-
-
-async def test_openai_responses_file_cache_point(allow_model_requests: None):
-    """Pin breakpoint translation for the remaining supported Responses content type."""
-    c = response_message(
-        [
-            ResponseOutputMessage(
-                id='output-1',
-                content=cast(list[Content], [ResponseOutputText(text='done', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ]
-    )
-    mock_client = MockOpenAIResponses.create_mock(c)
-    model = OpenAIResponsesModel('gpt-5.6-sol', provider=OpenAIProvider(openai_client=mock_client))
-
-    await Agent(model).run(
-        [BinaryContent(b'%PDF-1.4', media_type='application/pdf'), CachePoint(), 'Summarize the reference.']
-    )
-
-    request_input = get_mock_responses_kwargs(mock_client)[0]['input']
-    content = request_input[0]['content']
-    assert isinstance(content, list)
-    first_content = cast(dict[str, Any], content[0])
-    assert first_content['type'] == 'input_file'
-    assert first_content.get('prompt_cache_breakpoint') == {'mode': 'explicit'}
-
-
-async def test_openrouter_responses_image_cache_point_raises(allow_model_requests: None):
-    """Not a VCR test: OpenRouter documents breakpoints only on `input_text` blocks."""
-    mock_client = MockOpenAIResponses.create_mock(response_message([]))
-    model = OpenAIResponsesModel('openai/gpt-5.6-sol', provider=OpenRouterProvider(openai_client=mock_client))
-
-    with pytest.raises(UserError, match="cannot follow an OpenAI 'input_image' content block"):
-        await Agent(model).run([ImageUrl('https://example.com/reference.png'), CachePoint(), 'Describe it.'])
-
-    assert get_mock_responses_kwargs(mock_client) == []
-
-
-async def test_openai_responses_prompt_cache_options_not_sent_to_unsupported_model(
-    allow_model_requests: None,
-):
-    """Not a VCR test: omission must hold before the SDK serializes or sends a request."""
-    c = response_message(
-        [
-            ResponseOutputMessage(
-                id='output-1',
-                content=cast(list[Content], [ResponseOutputText(text='done', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ]
-    )
-    mock_client = AsyncOpenAI(api_key='test-key')
-    create = AsyncMock(return_value=c)
-    mock_client.responses.create = create
-    model = OpenAIResponsesModel('gpt-4o', provider=OpenAIProvider(openai_client=mock_client))
-    settings = OpenAIResponsesModelSettings(openai_prompt_cache_options={'mode': 'explicit', 'ttl': '30m'})
-
-    result = await Agent(model, model_settings=settings).run(['Stable context.', CachePoint(), 'Use it.'])
-
-    assert result.output == 'done'
-    assert create.await_count == 1
-    request = create.call_args.kwargs
-    assert 'prompt_cache_options' not in request
-    assert request['input'] == [
-        {
-            'role': 'user',
-            'content': [
-                {'type': 'input_text', 'text': 'Stable context.'},
-                {'type': 'input_text', 'text': 'Use it.'},
-            ],
-        }
-    ]
-
-
-async def test_azure_responses_prompt_cache_fields_are_omitted(allow_model_requests: None):
-    c = response_message(
-        [
-            ResponseOutputMessage(
-                id='output-1',
-                content=cast(list[Content], [ResponseOutputText(text='done', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ]
-    )
-    mock_client = MockOpenAIResponses.create_mock(c)
-    model = OpenAIResponsesModel(
-        'gpt-5.6-sol', provider=AzureProvider(openai_client=cast(AsyncAzureOpenAI, mock_client))
-    )
-    settings = OpenAIResponsesModelSettings(openai_prompt_cache_options={'mode': 'explicit', 'ttl': '30m'})
-
-    await Agent(model, model_settings=settings).run(['Stable context.', CachePoint(), 'Use it.'])
-
-    request = get_mock_responses_kwargs(mock_client)[0]
-    assert 'prompt_cache_options' not in request
-    assert request['input'] == [
-        {
-            'role': 'user',
-            'content': [
-                {'type': 'input_text', 'text': 'Stable context.'},
-                {'type': 'input_text', 'text': 'Use it.'},
-            ],
-        }
-    ]
-
-
-async def test_azure_responses_prompt_cache_fields_are_omitted_with_openai_provider(allow_model_requests: None):
-    c = response_message(
-        [
-            ResponseOutputMessage(
-                id='output-1',
-                content=cast(list[Content], [ResponseOutputText(text='done', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ]
-    )
-    async with AsyncAzureOpenAI(
-        azure_endpoint='https://example-resource.openai.azure.com',
-        api_version='2026-07-01-preview',
-        api_key='test-key',
-    ) as client:
-        create = AsyncMock(return_value=c)
-        client.responses.create = create
-        model = OpenAIResponsesModel('gpt-5.6-sol', provider=OpenAIProvider(openai_client=client))
-        settings = OpenAIResponsesModelSettings(openai_prompt_cache_options={'mode': 'explicit', 'ttl': '30m'})
-
-        await Agent(model, model_settings=settings).run(['Stable context.', CachePoint(), 'Use it.'])
-
-    request = create.call_args.kwargs
-    assert 'prompt_cache_options' not in request
-    assert request['input'] == [
-        {
-            'role': 'user',
-            'content': [
-                {'type': 'input_text', 'text': 'Stable context.'},
-                {'type': 'input_text', 'text': 'Use it.'},
-            ],
-        }
-    ]
-
-
-async def test_openai_responses_cache_point_first_content_raises(allow_model_requests: None):
-    c = response_message([])
-    mock_client = MockOpenAIResponses.create_mock(c)
-    model = OpenAIResponsesModel('gpt-5.6-sol', provider=OpenAIProvider(openai_client=mock_client))
-
-    with pytest.raises(UserError, match='`CachePoint` cannot be the first item in an OpenAI user prompt'):
-        await Agent(model).run([CachePoint(), 'This should fail.'])
-
-    assert get_mock_responses_kwargs(mock_client) == []
 
 
 @pytest.mark.parametrize(
@@ -3801,105 +3424,6 @@ async def test_openai_responses_usage_without_tokens_details(allow_model_request
 
     assert result.usage == snapshot(
         RunUsage(input_tokens=14, output_tokens=1, details={'reasoning_tokens': 0}, requests=1)
-    )
-
-
-async def test_openai_responses_maps_cache_write_usage(allow_model_requests: None):
-    """Not a VCR test: a synthetic response isolates the internal usage-field mapping."""
-    c = response_message(
-        [
-            ResponseOutputMessage(
-                id='123',
-                content=cast(list[Content], [ResponseOutputText(text='4', type='output_text', annotations=[])]),
-                role='assistant',
-                status='completed',
-                type='message',
-            )
-        ],
-        usage=ResponseUsage(
-            input_tokens=2006,
-            input_tokens_details=InputTokensDetails(cached_tokens=1920, cache_write_tokens=64),
-            output_tokens=300,
-            output_tokens_details=OutputTokensDetails(reasoning_tokens=10),
-            total_tokens=2306,
-        ),
-    )
-    mock_client = MockOpenAIResponses.create_mock(c)
-    model = OpenAIResponsesModel('gpt-5.6-sol', provider=OpenAIProvider(openai_client=mock_client))
-
-    result = await Agent(model=model).run('What is 2+2?')
-
-    assert result.usage == RunUsage(
-        requests=1,
-        input_tokens=2006,
-        cache_write_tokens=64,
-        cache_read_tokens=1920,
-        output_tokens=300,
-        details={'reasoning_tokens': 10},
-    )
-
-
-async def test_openai_responses_stream_maps_cache_write_usage(allow_model_requests: None):
-    """Not a VCR test: synthetic stream events isolate the internal usage-field mapping."""
-    base_response = resp.Response(
-        id='resp_001',
-        model='gpt-5.6-sol',
-        object='response',
-        created_at=1704067200,
-        output=[],
-        parallel_tool_calls=True,
-        tool_choice='auto',
-        tools=[],
-    )
-    response_usage = ResponseUsage(
-        input_tokens=2006,
-        input_tokens_details=InputTokensDetails(cached_tokens=1920, cache_write_tokens=64),
-        output_tokens=300,
-        output_tokens_details=OutputTokensDetails(reasoning_tokens=10),
-        total_tokens=2306,
-    )
-    stream: list[resp.ResponseStreamEvent] = [
-        resp.ResponseCreatedEvent(response=base_response, type='response.created', sequence_number=0),
-        resp.ResponseOutputItemAddedEvent(
-            item=ResponseOutputMessage(
-                id='msg_001',
-                content=[],
-                role='assistant',
-                status='in_progress',
-                type='message',
-            ),
-            output_index=0,
-            type='response.output_item.added',
-            sequence_number=1,
-        ),
-        resp.ResponseTextDeltaEvent(
-            item_id='msg_001',
-            output_index=0,
-            content_index=0,
-            delta='done',
-            logprobs=[],
-            type='response.output_text.delta',
-            sequence_number=2,
-        ),
-        resp.ResponseCompletedEvent(
-            response=base_response.model_copy(update={'status': 'completed', 'usage': response_usage}),
-            type='response.completed',
-            sequence_number=3,
-        ),
-    ]
-    mock_client = MockOpenAIResponses.create_mock_stream(stream)
-    model = OpenAIResponsesModel('gpt-5.6-sol', provider=OpenAIProvider(openai_client=mock_client))
-
-    async with Agent(model).run_stream('Solve this.') as result:
-        assert await result.get_output() == 'done'
-
-    assert result.usage == RunUsage(
-        requests=1,
-        input_tokens=2006,
-        cache_write_tokens=64,
-        cache_read_tokens=1920,
-        output_tokens=300,
-        details={'reasoning_tokens': 10},
     )
